feat(employee-schedule,accounting): OT decimal input + Customers split into 3 lists w/ pagination (v2.71.0)
- Employee Schedule: OT hours dropdown -> free numeric input (decimals like 1.5), numbers-only with red error on invalid; clamp 0-MAX_OT_HOURS, 2dp round. Schema EmpSchedule.otHours Int -> Float; route drops .int().
- Accounting Customers/Vendors: split into 3 tabs (Individual / Corporation / Vendors); Add Customer popup relabels Name -> Company Name when Corporation; PAGE_SIZE=12 pagination per tab.
- Docs: CLAUDE.md + ARCHITECTURE.md bumped to v2.71.0.

+++ backend/src/routes/employeeSchedule.ts
@@ -53,7 +53,7 @@ const CellBody = z.object({
   employeeId: z.string().uuid(),
   date: z.string().regex(DATE_RE),
   status: z.nativeEnum(AttendanceStatus).nullable(),
-  otHours: z.coerce.number().int().min(0).max(5).default(0),
+  otHours: z.coerce.number().min(0).max(5).default(0), // decimals allowed (e.g. 1.5)
 })
 
 const ReportQuery = z.object({
